# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harness daemon API",
    "description": "REST API served by the Harness daemon (harness-daemon). The API is dual-transport: most endpoints mirror a WebSocket JSON-RPC method, recorded per operation as the `x-websocket-method` extension. Generated from annotated handlers - do not edit docs/api/openapi.json by hand; regenerate with `mise run openapi:generate`.",
    "license": {
      "name": "MIT"
    },
    "version": "50.0.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}/task": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_create",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.create"
      }
    },
    "/v1/sessions/{session_id}/tasks/{task_id}": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.delete"
      }
    },
...
    "/v1/sessions/{session_id}/tasks/{task_id}/assign": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_assign",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskAssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task assigned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.assign"
      }
    },
    "/v1/sessions/{session_id}/tasks/{task_id}/checkpoint": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_checkpoint",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCheckpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task checkpoint recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.checkpoint"
      }
    },
...
    "/v1/sessions/{session_id}/tasks/{task_id}/drop": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_drop",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDropRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task dropped from the target's queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.drop"
      }
    },
    "/v1/sessions/{session_id}/tasks/{task_id}/queue-policy": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_queue_policy",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskQueuePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task queue policy updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.queue_policy"
      }
    },
...
    "/v1/sessions/{session_id}/tasks/{task_id}/status": {
      "post": {
        "tags": [
          "tasks"
        ],
        "operationId": "post_task_update",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "description": "Task identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Task status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "task.update"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DaemonErrorBody": {
        "type": "object",
        "description": "Error envelope returned by daemon handlers on failure.\n\nMirrors the dominant shape produced by `error_status_and_body`\n(`{\"error\": {\"code\", \"message\", \"details\"}}`). A few endpoints emit\nalternate ad-hoc error shapes for specific conditions; those are noted in\nthe relevant operation descriptions.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/DaemonErrorDetail"
          }
        }
      },
...
      "SessionDetail": {
        "type": "object",
        "required": [
          "session",
          "agents",
          "tasks",
          "signals",
          "agent_activity"
        ],
        "properties": {
          "agent_activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentToolActivitySummary"
            }
          },
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRegistrationWire"
            }
          },
          "observer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ObserverSummary"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          },
          "signals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSignalRecord"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            }
          }
        }
      },
...
      "TaskAssignRequest": {
        "type": "object",
        "required": [
          "actor",
          "agent_id"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "agent_id": {
            "type": "string"
          }
        }
      },
      "TaskCheckpointRequest": {
        "type": "object",
        "required": [
          "actor",
          "summary",
          "progress"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "progress": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 0
          },
          "summary": {
            "type": "string"
          }
        }
      },
...
      "TaskCreateRequest": {
        "type": "object",
        "required": [
          "actor",
          "title",
          "severity"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "context": {
            "type": [
              "string",
              "null"
            ]
          },
          "severity": {
            "$ref": "#/components/schemas/TaskSeverity"
          },
          "suggested_fix": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          }
        }
      },
      "TaskDeleteRequest": {
        "type": "object",
        "required": [
          "actor"
        ],
        "properties": {
          "actor": {
            "type": "string"
          }
        }
      },
...
      "TaskQueuePolicyRequest": {
        "type": "object",
        "required": [
          "actor",
          "queue_policy"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "queue_policy": {
            "$ref": "#/components/schemas/TaskQueuePolicy"
          }
        }
      },
...
    },
    "responses": {
      "RemoteAuthRequired": {
        "description": "Missing, invalid, or unauthorized daemon credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestUriTooLong": {
        "description": "Request URI exceeds the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestBodyTooLarge": {
        "description": "Request body exceeds the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestHeadersTooLarge": {
        "description": "Request headers exceed the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestThrottled": {
        "description": "Remote request or connection concurrency limit reached",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteServiceUnavailable": {
        "description": "Remote auth store or request limits unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestTimedOut": {
        "description": "Remote request exceeded the configured timeout",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      }
    }
  },
...
}